workers/switti/requirements.txt
redis
Pillow
accelerate
